Adds a check on all pull requests for baking time
After pull requests are updated, a 2 day waiting period is required to ensure plenty of time for additional reviews and comments.

This repositories handles changes that are sensitive to the OpenSearch-Project involving many maintainers/contributors by using a tool to enforce the 'bake time' it removes maintainers from having to remember how long to wait and communicates to contributors when a pull request is ready to merge.

The underlying github actions for this change are in https://github.com/peternied/bake-time to make this functionality reusable in other repositories.

Signed-off-by: Peter Nied <petern@amazon.com>